backend/items/urls.py
from rest_framework import routers
from django.urls import path, include
from .views import ItemViewSet

router = routers.DefaultRouter()
router.register(r'items', ItemViewSet)

urlpatterns = [
    path('', include(router.urls)),
]